marshal byte array as hex by default

+++ go_types.go
@@ -72,5 +72,5 @@ type goStringHeader struct {
 type goSliceHeader struct {
 	Data uintptr
 	Len  uintptr
-	Cap  int
+	Cap  uintptr
 }
